Refactor internals: Part 1

+++ test/symbols.jl
@@ -41,42 +41,42 @@
 
         @testset "DNA" begin
             # 2 bits
-            @test encode(DNAAlphabet{2}(), DNA_A) === 0x00
-            @test encode(DNAAlphabet{2}(), DNA_C) === 0x01
-            @test encode(DNAAlphabet{2}(), DNA_G) === 0x02
-            @test encode(DNAAlphabet{2}(), DNA_T) === 0x03
+            @test encode(DNAAlphabet{2}(), DNA_A) === UInt(0x00)
+            @test encode(DNAAlphabet{2}(), DNA_C) === UInt(0x01)
+            @test encode(DNAAlphabet{2}(), DNA_G) === UInt(0x02)
+            @test encode(DNAAlphabet{2}(), DNA_T) === UInt(0x03)
             @test_throws EncodeError encode(DNAAlphabet{2}(), DNA_M)
             @test_throws EncodeError encode(DNAAlphabet{2}(), DNA_N)
             @test_throws EncodeError encode(DNAAlphabet{2}(), DNA_Gap)
 
             # 4 bits
             for nt in BioSymbols.alphabet(DNA)
-                @test encode(DNAAlphabet{4}(), nt) === reinterpret(UInt8, nt)
+                @test encode(DNAAlphabet{4}(), nt) === UInt(reinterpret(UInt8, nt))
             end
             @test_throws EncodeError encode(DNAAlphabet{4}(), reinterpret(DNA, 0b10000))
         end
 
         @testset "RNA" begin
             # 2 bits
-            @test encode(RNAAlphabet{2}(), RNA_A) === 0x00
-            @test encode(RNAAlphabet{2}(), RNA_C) === 0x01
-            @test encode(RNAAlphabet{2}(), RNA_G) === 0x02
-            @test encode(RNAAlphabet{2}(), RNA_U) === 0x03
+            @test encode(RNAAlphabet{2}(), RNA_A) === UInt(0x00)
+            @test encode(RNAAlphabet{2}(), RNA_C) === UInt(0x01)
+            @test encode(RNAAlphabet{2}(), RNA_G) === UInt(0x02)
+            @test encode(RNAAlphabet{2}(), RNA_U) === UInt(0x03)
             @test_throws EncodeError encode(RNAAlphabet{2}(), RNA_M)
             @test_throws EncodeError encode(RNAAlphabet{2}(), RNA_N)
             @test_throws EncodeError encode(RNAAlphabet{2}(), RNA_Gap)
 
             # 4 bits
             for nt in BioSymbols.alphabet(RNA)
-                @test encode(RNAAlphabet{4}(), nt) === reinterpret(UInt8, nt)
+                @test encode(RNAAlphabet{4}(), nt) === UInt(reinterpret(UInt8, nt))
             end
             @test_throws EncodeError encode(RNAAlphabet{4}(), reinterpret(RNA, 0b10000))
         end
 
         @testset "AminoAcid" begin
-            @test encode(AminoAcidAlphabet(), AA_A) === 0x00
+            @test encode(AminoAcidAlphabet(), AA_A) === UInt(0x00)
             for aa in BioSymbols.alphabet(AminoAcid)
-                @test encode(AminoAcidAlphabet(), aa) === convert(UInt8, aa)
+                @test encode(AminoAcidAlphabet(), aa) === convert(UInt, aa)
             end
             @test_throws BioSequences.EncodeError encode(AminoAcidAlphabet(), BioSymbols.AA_INVALID)
         end
